Switch from "Adminpassword System" to user based system
This is still a WIP Project. Yes it works (otherwise I would not commit it). But I cant guarantee that there is no security issue, the token must be entered manually when installing/updating, there is only a admin user, the user can not change its password/any other value, there is no registration option and more. But this are big changes and the biggest step I made with this commit.